select the last completed week to look at
The last week as returned by the JF API can be one that just started with the end date in the future. We now ensure that for more predictable reporting we look at the last week with end date in the past. We will also look at due date changes from the firday in that week back to the Saturday before that to sync with shiproom schedule.